Backend/python-ml/requirements.txt
google-generativeai>=0.3.0
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
sqlalchemy>=1.4.0
python-dateutil>=2.8.0
types-python-dateutil>=2.8.0